tests for drag-drop-context
