+++ src/view/use-style-marshal/index.js
@@ -1,1 +1,2 @@
+// @flow
 export { default } from './use-style-marshal';
