+++ test/unit/view/drag-drop-context/unmount.spec.js
@@ -1,14 +1,11 @@
 // @flow
 import React from 'react';
 import { mount } from 'enzyme';
-import App from './app';
 import DragDropContext from '../../../../src/view/drag-drop-context';
 
 it('should not throw when unmounting', () => {
   const wrapper = mount(
-    <DragDropContext onDragEnd={() => {}}>
-      <App />
-    </DragDropContext>,
+    <DragDropContext onDragEnd={() => {}}>{null}</DragDropContext>,
   );
 
   expect(() => wrapper.unmount()).not.toThrow();
@@ -19,9 +16,7 @@ it('should clean up any window event handlers', () => {
   jest.spyOn(window, 'removeEventListener');
 
   const wrapper = mount(
-    <DragDropContext onDragEnd={() => {}}>
-      <App />
-    </DragDropContext>,
+    <DragDropContext onDragEnd={() => {}}>{null}</DragDropContext>,
   );
 
   wrapper.unmount();
